Make packet type a short rather than a byte.
It is conceivable that a user wants more than 256 types of packets.

+++ src/lucid/client/TcpConnection.java
@@ -50,7 +50,7 @@ public class TcpConnection implements Runnable {
 			channel.connect(new InetSocketAddress(host, port));
 			channel.setOption(StandardSocketOptions.TCP_NODELAY, true);
 			
-			Packet packet = new Packet(0);
+			Packet packet = new Packet((short) 0);
 			packet.addLong(UniqueGenerator.unique);
 			send(packet);
 			read();
